Implement Get and List methods
This driver, as it was implemented a few months ago, relied on an
unfinished version of Docker volume plugin API (ironically it was
finished just a few weeks after I committed the code, see [1]).

In particular, there were no Get() and List() methods, and
the semantics was a bit different, too (as far as I remember,
Create() was called if you try to start a container with a
non-existing volume).

The documentation on the Docker volume plugin API is sparse,
but it looks like Get() is used to check whether the volume
is present, otherwise it's very similar to Path().

List() returns the list of volumes known to the plugin.

[1] https://github.com/docker/go-plugins-helpers/commit/5424fab62fa

Signed-off-by: Kir Kolyshkin <kir@openvz.org>