Create(): rm volume dir on failure
If ploop create failed, we need to clean things up. If not,
we'll not be able to create a volume under that name later,
as Mkdir() will return EEXIST (which we don't want to ignore).

Note this is a great example of conference driven development;
I discovered and fixed the bug right here at the DockerCon 2016,
Virtuozzo booth (S28) while showing it to a colleague.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>